+++ config/webpack/webpack.config.js
@@ -8,10 +8,12 @@ const CleanPlugin = require('clean-webpack-plugin');
 const TerserPlugin = require('terser-webpack-plugin');
 const browserslist = require('browserslist-config-terra');
 const merge = require('webpack-merge');
+const DuplicatePackageCheckerPlugin = require('duplicate-package-checker-webpack-plugin');
 const aggregateTranslations = require('../../scripts/aggregate-translations/aggregate-translations');
+const ThemeAggregator = require('../../scripts/aggregate-themes/theme-aggregator');
 
 const webpackConfig = (options, env, argv) => {
-  const { rootPath, resolveModules } = options;
+  const { rootPath, resolveModules, themeFile } = options;
 
   const production = argv.p;
   let filename = production ? '[name]-[chunkhash]' : '[name]';
@@ -24,6 +26,7 @@ const webpackConfig = (options, env, argv) => {
     entry: {
       raf: 'raf/polyfill',
       'babel-polyfill': 'babel-polyfill',
+      ...themeFile && { theme: themeFile },
     },
     module: {
       rules: [
@@ -83,6 +86,9 @@ const webpackConfig = (options, env, argv) => {
           PostCSSCustomProperties({ preserve: true }),
         ],
       }),
+      new DuplicatePackageCheckerPlugin({
+        showHelp: false,
+      }),
     ],
     resolve: {
       extensions: ['.js', '.jsx'],
@@ -136,7 +142,7 @@ const webpackConfig = (options, env, argv) => {
 };
 
 const defaultWebpackConfig = (env = {}, argv = {}) => {
-  const disableAggregateTranslations = env.disableAggregateTranslations;
+  const { disableAggregateTranslations } = env;
 
   const processPath = process.cwd();
   /* Get the root path of a mono-repo process call */
@@ -148,7 +154,9 @@ const defaultWebpackConfig = (env = {}, argv = {}) => {
     resolveModules.unshift(path.resolve(rootPath, 'aggregated-translations'));
   }
 
-  const options = { rootPath, resolveModules };
+  const themeFile = ThemeAggregator.aggregate();
+
+  const options = { rootPath, resolveModules, themeFile };
 
   return webpackConfig(options, env, argv);
 };