+++ src/wdio/services/ServeStaticService.js
@@ -1,16 +1,16 @@
 import serveStatic from '../../../scripts/serve/serve-static';
 import SERVICE_DEFAULTS from '../../../config/wdio/services.default-config';
+import Logger from '../../../scripts/utils/logger';
 
 const { serveStatic: SERVE_STATIC_DEFAULTS } = SERVICE_DEFAULTS;
+const context = '[Terra-Toolkit:serve-static-service]';
 
 export default class ServeStaticService {
   async onPrepare(config = {}) {
-    const site = config.site;
-    const webpackConfig = config.webpackConfig;
+    const { site, webpackConfig, locale } = config;
 
     if (!webpackConfig && !site) {
-      // eslint-disable-next-line no-console
-      console.warn('[Terra-Toolkit:serve-static] No webpack configuration provided');
+      Logger.warn('No webpack configuration provided', { context });
       return;
     }
 
@@ -18,7 +18,6 @@ export default class ServeStaticService {
     const port = (config.serveStatic || {}).port || SERVE_STATIC_DEFAULTS.port;
     const index = (config.serveStatic || {}).index || SERVE_STATIC_DEFAULTS.index;
     // Explicitly not providing a fallback locale. Providing a fallback will lock the locale for all test runs when using the tt-wdio-runner.
-    const locale = (config || {}).locale;
 
     // Ensure the server was properly shut down.
     if (this.server) {
@@ -52,8 +51,7 @@ export default class ServeStaticService {
 
   stop() {
     return new Promise((resolve) => {
-      // eslint-disable-next-line no-console
-      console.log('[Terra-Toolkit:serve-static] Closing Server');
+      Logger.log('Closing Server', { context });
       if (this.server) {
         this.server.close();
         this.server = null;